+++ cpp/QmlInterface.cpp
@@ -220,4 +220,28 @@ CMUserInformation QmlInterface::cmUserInformation(){
     return _cmUserInformation;
 }
 
+QString QmlInterface::userId(){
+    return _userId;
+}
+
+void QmlInterface::setUserId(const QString userID){
+    if(userID!=_userId){
+        qInfo()<<"Default user ID set to " << userID;
+        _userId = userID;
+        emit userIdChanged(_userId);
+    }
+}
+
+QString QmlInterface::getServerAddress(){
+    return _serverAddress;
+};
+
+void QmlInterface::setServerAddress(const QString &server){
+    if(server!=_serverAddress){
+        qInfo()<<"Default server set to " << server;
+        _serverAddress = server;
+        emit serverAddressChanged(_serverAddress);
+    }
+};
+
 }