+++ cpp/QmlInterface.h
@@ -10,11 +10,10 @@
 #include <matrix-client-library/voip/CallManager.h>
 #include <matrix-client-library/UserSettings.h>
 #include <matrix-client-library/voip/WebRTCSession.h>
-#include <matrix-client-library/CMUserInfo.h>
 #include "RoomListModel.h"
 #include "RoomListItem.h"
 
-namespace PX::GUI::MATRIX{
+namespace PX::GUI::MATRIX {
 class QmlInterface : public QObject {
     Q_OBJECT
 public: 
@@ -29,16 +28,18 @@ public:
     void setAutoAcceptCall(bool mode) { _callAutoAccept = mode; };
     bool autoAcceptCall() { return _callAutoAccept; };
     
+signals:
+    void userIdChanged(const QString &userId);
+    void serverAddressChanged(const QString &server);
+
 public slots:
     virtual void setVideoCallItem() = 0;
-    QString getServerAddress(){return _serverAddress;};
-    void setServerAddress(QString server){
-        qInfo()<<"Server default set to "<<server;
-        _serverAddress = server;
-    };
+    QString userId();
+    void setUserId(const QString userID);
+    QString getServerAddress();
+    void setServerAddress(const QString &server);
     void setCMUserInformation(const CMUserInformation &info);
     CMUserInformation cmUserInformation();
-
 
 private slots:
     void initiateFinishedCB();
@@ -46,15 +47,16 @@ private slots:
 
 private:
     void checkCacheDirectory();
-    
+
     bool _callAutoAccept = false;
     RoomListModel *_roomListModel = nullptr;
-    Client *_client = nullptr;
+    Client      *_client = nullptr;
     CallManager *_callMgr = nullptr;
     VerificationManager *_verificationManager;
     QSharedPointer<UserSettings> _userSettings;
     QString _defaultUserIdFormat = "@user:matrix.org";
     QString _serverAddress = "";
+    QString _userId = "";
     CMUserInformation _cmUserInformation;
 };
 }